feat: AI 분석 기반 제품 추천 기능

+++ backend/src/main/java/com/yunyun/meallog/global/jwt/JwtUtil.java
@@ -47,9 +47,9 @@ public class JwtUtil {
 //    }
 
     // [수정]
-    public Integer getUserId(String token) {
+    public Long getUserId(String token) {
         Claims claims = getClaims(token);
-        return Integer.parseInt(claims.get("id").toString()); // [수정]
+        return Long.parseLong(claims.get("id").toString()); // [수정]
     }
