+++ backend/src/main/java/com/yunyun/meallog/user/controller/UserController.java
@@ -29,14 +29,14 @@ public class UserController {
      * 회원가입 시 추가적인 정보를 함께 반환해야 할 경우 Map에 새로운 키-값 쌍을 추가하여 확장 가능
      */
     @PostMapping("/signup")
-    public ResponseEntity<Map<String, Integer>> signUp(@Valid @RequestBody SignupStep1RequestDto requestDto) {
-        Integer userId = userService.signUp(requestDto);
-        Map<String, Integer> response = Collections.singletonMap("userId", userId);
+    public ResponseEntity<Map<String, Long>> signUp(@Valid @RequestBody SignupStep1RequestDto requestDto) {
+        Long userId = userService.signUp(requestDto);
+        Map<String, Long> response = Collections.singletonMap("userId", userId);
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
     @PutMapping("/{userId}/profile")
-    public ResponseEntity<String> updateProfile(@PathVariable Integer userId, @RequestBody SignupStep2RequestDto requestDto) {
+    public ResponseEntity<String> updateProfile(@PathVariable Long userId, @RequestBody SignupStep2RequestDto requestDto) {
         userService.updateProfile(userId, requestDto);
         return ResponseEntity.ok("프로필이 성공적으로 업데이트되었습니다.");
     }
@@ -66,7 +66,7 @@ public class UserController {
      */
     @GetMapping("/me")
     public ResponseEntity<UserResponseDto> getMyProfile(HttpServletRequest request) {
-        Integer userId = (Integer) request.getAttribute("userId");
+        Long userId = (Long) request.getAttribute("userId");
         UserResponseDto response = userService.getMyProfile(userId);
         return ResponseEntity.ok(response);
     }
