+++ backend/src/main/java/com/yunyun/meallog/user/service/impl/UserServiceImpl.java
@@ -28,7 +28,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public Integer signUp(SignupStep1RequestDto requestDto) {
+    public Long signUp(SignupStep1RequestDto requestDto) {
         // 이메일 중복 체크
         userDao.findByEmail(requestDto.getEmail())
                 .ifPresent(user -> { throw new CustomException(ErrorCode.DUPLICATE_EMAIL); });
@@ -63,7 +63,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public void updateProfile(Integer userId, SignupStep2RequestDto requestDto) {
+    public void updateProfile(Long userId, SignupStep2RequestDto requestDto) {
         User userToUpdate = userDao.findById(userId)
                 .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
 
@@ -103,7 +103,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserResponseDto getMyProfile(Integer userId) {
+    public UserResponseDto getMyProfile(Long userId) {
         User user = userDao.findById(userId)
                 .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
         return UserResponseDto.from(user);
